Denormalize base_cli.rb test class again
